feat: fix warning while migrate on tb_data_perusahaan and tb_pencapaian_ekspor

+++ app/Database/Migrations/2024-10-03-071049_CreateTbDataPerusahaan.php
@@ -84,6 +84,8 @@ class CreateTbDataPerusahaan extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('tb_data_perusahaan');
+        if (!empty($result)) {
+            $this->forge->dropTable('tb_data_perusahaan');
+        }
     }
 }
