+++ app/Database/Migrations/2024-10-03-071237_CreateTbPencapaianEkspor.php
@@ -61,6 +61,10 @@ class CreateTbPencapaianEkspor extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('tb_pencapaian_ekspor');
+        $result = $this->db->query("SHOW TABLES LIKE 'tb_pencapaian_ekspor'")->getResult();
+
+        if (!empty($result)) {
+            $this->forge->dropTable('tb_pencapaian_ekspor');
+        }
     }
 }
